Fixes button layout issue in application group edit and authorization create pages; Renames Global Settings to Settings
