+++ app/settings/authorization/authorizationcreate.html
@@ -100,7 +100,7 @@
                             </div>
                         </div>
                         <div class="ui row">
-                            <div class="ui right aligned sixteen wide column">
+                            <div class="ui right aligned six wide column">
                                 <div class="buttonRow">
                                     <button type="button" class="ui basic button" (click)="cancelCreating()">
                                         Cancel
