+++ app/settings/settingsmenu.html
@@ -1,7 +1,7 @@
 <div class="ui sixteen column grid">
     <div class="ui row pageHeader">
         <div class="left aligned eight wide column">
-            <div class="content pageTitle">Global Settings</div>
+            <div class="content pageTitle">Settings</div>
         </div>
         <div class="right aligned eight wide column">&nbsp;</div>
     </div>
